+++ src/domain/email/args.rs
@@ -4,12 +4,12 @@
 //! arguments related to the email domain.
 
 use anyhow::Result;
-use clap::{App, Arg, ArgMatches, SubCommand};
+use clap::{Arg, ArgAction, ArgMatches, Command};
 
 use crate::{email, flag, folder, tpl, ui::table};
 
 const ARG_CRITERIA: &str = "criterion";
-const ARG_HEADERS: &str = "header";
+const ARG_HEADERS: &str = "headers";
 const ARG_ID: &str = "id";
 const ARG_MIME_TYPE: &str = "mime-type";
 const ARG_PAGE: &str = "page";
@@ -41,7 +41,7 @@ pub type Page = usize;
 pub type PageSize = usize;
 pub type Query = String;
 pub type Raw = bool;
-pub type RawEmail<'a> = &'a str;
+pub type RawEmail = String;
 pub type Sanitize = bool;
 pub type TextMime<'a> = &'a str;
 
@@ -57,9 +57,9 @@ pub enum Cmd<'a> {
     Move(Id<'a>, Folder<'a>),
     Read(Id<'a>, TextMime<'a>, Sanitize, Raw, Headers<'a>),
     Reply(Id<'a>, All, tpl::args::Headers<'a>, tpl::args::Body<'a>),
-    Save(RawEmail<'a>),
+    Save(RawEmail),
     Search(Query, table::args::MaxTableWidth, Option<PageSize>, Page),
-    Send(RawEmail<'a>),
+    Send(RawEmail),
     Sort(
         Criteria,
         Query,
@@ -145,73 +145,73 @@ pub fn matches<'a>(m: &'a ArgMatches) -> Result<Option<Cmd<'a>>> {
 }
 
 /// Represents the email subcommands.
-pub fn subcmds<'a>() -> Vec<App<'a, 'a>> {
+pub fn subcmds() -> Vec<Command> {
     vec![
         flag::args::subcmds(),
         tpl::args::subcmds(),
         vec![
-            SubCommand::with_name(CMD_ATTACHMENTS)
-                .aliases(&["attachment", "attach", "att", "at", "a"])
+            Command::new(CMD_ATTACHMENTS)
                 .about("Downloads all attachments of the targeted email")
                 .arg(email::args::id_arg()),
-            SubCommand::with_name(CMD_LIST)
-                .aliases(&["lst", "l"])
+            Command::new(CMD_LIST)
+                .alias("lst")
                 .about("Lists all emails")
                 .arg(page_size_arg())
                 .arg(page_arg())
                 .arg(table::args::max_width()),
-            SubCommand::with_name(CMD_SEARCH)
-                .aliases(&["s", "query", "q"])
+            Command::new(CMD_SEARCH)
+                .aliases(["query", "q"])
                 .about("Lists emails matching the given query")
                 .arg(page_size_arg())
                 .arg(page_arg())
                 .arg(table::args::max_width())
                 .arg(query_arg()),
-            SubCommand::with_name(CMD_SORT)
+            Command::new(CMD_SORT)
                 .about("Sorts emails by the given criteria and matching the given query")
                 .arg(page_size_arg())
                 .arg(page_arg())
                 .arg(table::args::max_width())
                 .arg(criteria_arg())
                 .arg(query_arg()),
-            SubCommand::with_name(CMD_WRITE)
+            Command::new(CMD_WRITE)
                 .about("Writes a new email")
-                .aliases(&["w", "new", "n"])
-                .args(&tpl::args::args()),
-            SubCommand::with_name(CMD_SEND)
+                .aliases(["new", "n"])
+                .args(tpl::args::args()),
+            Command::new(CMD_SEND)
                 .about("Sends a raw email")
                 .arg(raw_arg()),
-            SubCommand::with_name(CMD_SAVE)
+            Command::new(CMD_SAVE)
                 .about("Saves a raw email")
                 .arg(raw_arg()),
-            SubCommand::with_name(CMD_READ)
+            Command::new(CMD_READ)
                 .about("Reads text bodies of an email")
-                .arg(id_arg())
                 .arg(mime_type_arg())
                 .arg(sanitize_flag())
                 .arg(raw_flag())
-                .arg(headers_arg()),
-            SubCommand::with_name(CMD_REPLY)
-                .aliases(&["rep", "r"])
-                .about("Answers to an email")
-                .arg(id_arg())
-                .arg(reply_all_flag()),
-            SubCommand::with_name(CMD_FORWARD)
-                .aliases(&["fwd", "f"])
-                .about("Forwards an email")
+                .arg(headers_arg())
                 .arg(id_arg()),
-            SubCommand::with_name(CMD_COPY)
-                .aliases(&["cp", "c"])
+            Command::new(CMD_REPLY)
+                .about("Answers to an email")
+                .arg(reply_all_flag())
+                .args(tpl::args::args())
+                .arg(id_arg()),
+            Command::new(CMD_FORWARD)
+                .aliases(["fwd", "f"])
+                .about("Forwards an email")
+                .args(tpl::args::args())
+                .arg(id_arg()),
+            Command::new(CMD_COPY)
+                .alias("cp")
                 .about("Copies an email to the targeted folder")
                 .arg(id_arg())
                 .arg(folder::args::target_arg()),
-            SubCommand::with_name(CMD_MOVE)
-                .aliases(&["mv"])
+            Command::new(CMD_MOVE)
+                .alias("mv")
                 .about("Moves an email to the targeted folder")
                 .arg(id_arg())
                 .arg(folder::args::target_arg()),
-            SubCommand::with_name(CMD_DELETE)
-                .aliases(&["del", "d", "remove", "rm"])
+            Command::new(CMD_DELETE)
+                .aliases(["remove", "rm"])
                 .about("Deletes an email")
                 .arg(id_arg()),
         ],
@@ -220,29 +220,27 @@ pub fn subcmds<'a>() -> Vec<App<'a, 'a>> {
 }
 
 /// Represents the email id argument.
-pub fn id_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_ID)
+pub fn id_arg() -> Arg {
+    Arg::new(ARG_ID)
         .help("Specifies the target email")
         .value_name("ID")
         .required(true)
 }
 
 /// Represents the email id argument parser.
-pub fn parse_id_arg<'a>(matches: &'a ArgMatches<'a>) -> &'a str {
-    matches.value_of(ARG_ID).unwrap()
+pub fn parse_id_arg(matches: &ArgMatches) -> &str {
+    matches.get_one::<String>(ARG_ID).unwrap()
 }
 
 /// Represents the email sort criteria argument.
-pub fn criteria_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_CRITERIA)
-        .long("criterion")
-        .short("c")
+pub fn criteria_arg<'a>() -> Arg {
+    Arg::new(ARG_CRITERIA)
         .help("Email sorting preferences")
+        .long("criterion")
+        .short('c')
         .value_name("CRITERION:ORDER")
-        .takes_value(true)
-        .multiple(true)
-        .required(true)
-        .possible_values(&[
+        .action(ArgAction::Append)
+        .value_parser([
             "arrival",
             "arrival:asc",
             "arrival:desc",
@@ -268,55 +266,59 @@ pub fn criteria_arg<'a>() -> Arg<'a, 'a> {
 }
 
 /// Represents the email sort criteria argument parser.
-pub fn parse_criteria_arg<'a>(matches: &'a ArgMatches<'a>) -> String {
+pub fn parse_criteria_arg(matches: &ArgMatches) -> String {
     matches
-        .values_of(ARG_CRITERIA)
+        .get_many::<String>(ARG_CRITERIA)
         .unwrap_or_default()
+        .map(ToOwned::to_owned)
         .collect::<Vec<_>>()
         .join(" ")
 }
 
 /// Represents the email reply all argument.
-pub fn reply_all_flag<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_REPLY_ALL)
+pub fn reply_all_flag() -> Arg {
+    Arg::new(ARG_REPLY_ALL)
         .help("Includes all recipients")
-        .short("A")
         .long("all")
+        .short('a')
+        .action(ArgAction::SetTrue)
 }
 
 /// Represents the email reply all argument parser.
-pub fn parse_reply_all_flag<'a>(matches: &'a ArgMatches<'a>) -> bool {
-    matches.is_present(ARG_REPLY_ALL)
+pub fn parse_reply_all_flag(matches: &ArgMatches) -> bool {
+    matches.get_flag(ARG_REPLY_ALL)
 }
 
 /// Represents the page size argument.
-fn page_size_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_PAGE_SIZE)
+fn page_size_arg() -> Arg {
+    Arg::new(ARG_PAGE_SIZE)
         .help("Page size")
-        .short("s")
-        .long("size")
+        .long("page-size")
+        .short('s')
         .value_name("INT")
 }
 
 /// Represents the page size argument parser.
-fn parse_page_size_arg<'a>(matches: &'a ArgMatches<'a>) -> Option<usize> {
-    matches.value_of(ARG_PAGE_SIZE).and_then(|s| s.parse().ok())
+fn parse_page_size_arg(matches: &ArgMatches) -> Option<usize> {
+    matches
+        .get_one::<String>(ARG_PAGE_SIZE)
+        .and_then(|s| s.parse().ok())
 }
 
 /// Represents the page argument.
-fn page_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_PAGE)
+fn page_arg() -> Arg {
+    Arg::new(ARG_PAGE)
         .help("Page number")
-        .short("p")
+        .short('p')
         .long("page")
         .value_name("INT")
         .default_value("1")
 }
 
 /// Represents the page argument parser.
-fn parse_page_arg<'a>(matches: &'a ArgMatches<'a>) -> usize {
+fn parse_page_arg(matches: &ArgMatches) -> usize {
     matches
-        .value_of(ARG_PAGE)
+        .get_one::<String>(ARG_PAGE)
         .unwrap()
         .parse()
         .ok()
@@ -325,88 +327,93 @@ fn parse_page_arg<'a>(matches: &'a ArgMatches<'a>) -> usize {
 }
 
 /// Represents the email headers argument.
-pub fn headers_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_HEADERS)
+pub fn headers_arg() -> Arg {
+    Arg::new(ARG_HEADERS)
         .help("Shows additional headers with the email")
-        .short("h")
         .long("header")
+        .short('H')
         .value_name("STRING")
-        .multiple(true)
+        .action(ArgAction::Append)
 }
 
 /// Represents the email headers argument parser.
-pub fn parse_headers_arg<'a>(matches: &'a ArgMatches<'a>) -> Vec<&'a str> {
-    matches.values_of(ARG_HEADERS).unwrap_or_default().collect()
+pub fn parse_headers_arg(m: &ArgMatches) -> Vec<&str> {
+    m.get_many::<String>(ARG_HEADERS)
+        .unwrap_or_default()
+        .map(String::as_str)
+        .collect::<Vec<_>>()
 }
 
 /// Represents the sanitize flag.
-pub fn sanitize_flag<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_SANITIZE)
+pub fn sanitize_flag() -> Arg {
+    Arg::new(ARG_SANITIZE)
         .help("Sanitizes text bodies")
         .long("sanitize")
-        .short("s")
+        .short('s')
+        .action(ArgAction::SetTrue)
 }
 
 /// Represents the raw flag.
-pub fn raw_flag<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_RAW)
+pub fn raw_flag() -> Arg {
+    Arg::new(ARG_RAW)
         .help("Returns raw version of email")
         .long("raw")
-        .short("r")
+        .short('r')
+        .action(ArgAction::SetTrue)
 }
 
 /// Represents the sanitize flag parser.
-pub fn parse_sanitize_flag<'a>(matches: &'a ArgMatches<'a>) -> bool {
-    matches.is_present(ARG_SANITIZE)
+pub fn parse_sanitize_flag(m: &ArgMatches) -> bool {
+    m.get_flag(ARG_SANITIZE)
 }
 
 /// Represents the raw flag parser.
-pub fn parse_raw_flag<'a>(matches: &'a ArgMatches<'a>) -> bool {
-    matches.is_present(ARG_RAW)
+pub fn parse_raw_flag(m: &ArgMatches) -> bool {
+    m.get_flag(ARG_RAW)
 }
 
 /// Represents the email raw argument.
-pub fn raw_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_RAW).raw(true)
+pub fn raw_arg() -> Arg {
+    Arg::new(ARG_RAW).raw(true)
 }
 
 /// Represents the email raw argument parser.
-pub fn parse_raw_arg<'a>(matches: &'a ArgMatches<'a>) -> &'a str {
-    matches.value_of(ARG_RAW).unwrap_or_default()
+pub fn parse_raw_arg(m: &ArgMatches) -> String {
+    m.get_one::<String>(ARG_RAW).cloned().unwrap_or_default()
 }
 
 /// Represents the email MIME type argument.
-pub fn mime_type_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_MIME_TYPE)
+pub fn mime_type_arg() -> Arg {
+    Arg::new(ARG_MIME_TYPE)
         .help("MIME type to use")
-        .short("t")
+        .short('t')
         .long("mime-type")
         .value_name("MIME")
-        .possible_values(&["plain", "html"])
+        .value_parser(["plain", "html"])
         .default_value("plain")
 }
 
 /// Represents the email MIME type argument parser.
-pub fn parse_mime_type_arg<'a>(matches: &'a ArgMatches<'a>) -> &'a str {
-    matches.value_of(ARG_MIME_TYPE).unwrap()
+pub fn parse_mime_type_arg(matches: &ArgMatches) -> &str {
+    matches.get_one::<String>(ARG_MIME_TYPE).unwrap()
 }
 
 /// Represents the email query argument.
-pub fn query_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_QUERY)
+pub fn query_arg() -> Arg {
+    Arg::new(ARG_QUERY)
         .long_help("The query system depends on the backend, see the wiki for more details")
         .value_name("QUERY")
-        .multiple(true)
+        .num_args(1..)
         .required(true)
 }
 
 /// Represents the email query argument parser.
-pub fn parse_query_arg<'a>(matches: &'a ArgMatches<'a>) -> String {
+pub fn parse_query_arg(matches: &ArgMatches) -> String {
     matches
-        .values_of(ARG_QUERY)
+        .get_many::<String>(ARG_QUERY)
         .unwrap_or_default()
         .fold((false, vec![]), |(escape, mut cmds), cmd| {
-            match (cmd, escape) {
+            match (cmd.as_str(), escape) {
                 // Next command is an arg and needs to be escaped
                 ("subject", _) | ("body", _) | ("text", _) => {
                     cmds.push(cmd.to_string());