+++ src/domain/flag/args.rs
@@ -4,7 +4,7 @@
 //! arguments related to the email flag domain.
 
 use anyhow::Result;
-use clap::{App, AppSettings, Arg, ArgMatches, SubCommand};
+use clap::{Arg, ArgMatches, Command};
 use log::{debug, info};
 
 use crate::email;
@@ -15,7 +15,7 @@ const CMD_ADD: &str = "add";
 const CMD_REMOVE: &str = "remove";
 const CMD_SET: &str = "set";
 
-pub(crate) const CMD_FLAG: &str = "flag";
+pub(crate) const CMD_FLAG: &str = "flags";
 
 type Flags = String;
 
@@ -52,28 +52,27 @@ pub fn matches<'a>(m: &'a ArgMatches) -> Result<Option<Cmd<'a>>> {
 }
 
 /// Represents the flag subcommands.
-pub fn subcmds<'a>() -> Vec<App<'a, 'a>> {
-    vec![SubCommand::with_name(CMD_FLAG)
-        .aliases(&["flags", "flg"])
+pub fn subcmds<'a>() -> Vec<Command> {
+    vec![Command::new(CMD_FLAG)
         .about("Handles email flags")
-        .setting(AppSettings::SubcommandRequiredElseHelp)
+        .subcommand_required(true)
+        .arg_required_else_help(true)
         .subcommand(
-            SubCommand::with_name(CMD_ADD)
-                .aliases(&["a"])
+            Command::new(CMD_ADD)
                 .about("Adds flags to an email")
                 .arg(email::args::id_arg())
                 .arg(flags_arg()),
         )
         .subcommand(
-            SubCommand::with_name(CMD_REMOVE)
-                .aliases(&["rem", "rm", "r", "delete", "del", "d"])
+            Command::new(CMD_REMOVE)
+                .aliases(["delete", "del", "d"])
                 .about("Removes flags from an email")
                 .arg(email::args::id_arg())
                 .arg(flags_arg()),
         )
         .subcommand(
-            SubCommand::with_name(CMD_SET)
-                .aliases(&["s", "change", "c"])
+            Command::new(CMD_SET)
+                .aliases(["change", "c"])
                 .about("Sets flags of an email")
                 .arg(email::args::id_arg())
                 .arg(flags_arg()),
@@ -81,18 +80,19 @@ pub fn subcmds<'a>() -> Vec<App<'a, 'a>> {
 }
 
 /// Represents the flags argument.
-pub fn flags_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_FLAGS)
+pub fn flags_arg() -> Arg {
+    Arg::new(ARG_FLAGS)
         .value_name("FLAGS…")
-        .multiple(true)
+        .num_args(1..)
         .required(true)
 }
 
 /// Represents the flags argument parser.
-pub fn parse_flags_arg<'a>(matches: &'a ArgMatches<'a>) -> String {
+pub fn parse_flags_arg(matches: &ArgMatches) -> String {
     matches
-        .values_of(ARG_FLAGS)
+        .get_many::<String>(ARG_FLAGS)
         .unwrap_or_default()
+        .cloned()
         .collect::<Vec<_>>()
         .join(" ")
 }